+++ packages/bundler/src/submitter.ts
@@ -4,16 +4,13 @@ import {
   Address,
   calculateOperationDigest,
   ProvenOperation,
+  OperationStatus,
 } from "@nocturne-xyz/sdk";
 import { parseEventsFromContractReceipt } from "@nocturne-xyz/sdk/dist/src/sdk/utils/ethers";
 import { Job, Worker } from "bullmq";
 import IORedis from "ioredis";
 import { ethers, providers } from "ethers";
-import {
-  OperationStatus,
-  OPERATION_BATCH_QUEUE,
-  OperationBatchJobData,
-} from "./common";
+import { OPERATION_BATCH_QUEUE, OperationBatchJobData } from "./common";
 import { getRedis } from "./utils";
 import { StatusDB } from "./db";
 import * as JSON from "bigint-json-serialization";